requirements.txt
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
defusedxml==0.7.1
fonttools==4.54.1
fpdf2==2.8.1
idna==3.10
pillow==10.4.0
requests==2.32.3
soupsieve==2.6
urllib3==2.2.3
